Move validated-ctx types to routes root
