+++ src/routes/auth/accept-invite/schema.ts
@@ -1,6 +1,6 @@
 import { z } from 'zod'
 
-import type { ValidatedJsonCtx } from '../../@shared'
+import type { ValidatedJsonCtx } from '@/routes/validated-ctx'
 
 import { requestValidationRules as rules } from '../../@shared'
