+++ src/routes/auth/logout/handler.ts
@@ -1,7 +1,7 @@
+import type { AppCtx } from '@/routes/validated-ctx'
+
 import { deleteRefreshCookie, getRefreshCookie, HttpStatus } from '@/net/http'
 import { logout } from '@/use-cases/auth/logout'
-
-import type { AppCtx } from '../../@shared'
 
 export const logoutHandler = async (c: AppCtx) => {
   const refreshToken = getRefreshCookie(c)
