+++ src/routes/auth/refresh-token/handler.ts
@@ -1,7 +1,7 @@
+import type { AppCtx } from '@/routes/validated-ctx'
+
 import { getDeviceInfo, getRefreshCookie, HttpStatus, setRefreshCookie } from '@/net/http'
 import { refreshAuthTokens } from '@/use-cases/auth'
-
-import type { AppCtx } from '../../@shared'
 
 export const refreshTokenHandler = async (c: AppCtx) => {
   const refreshToken = getRefreshCookie(c)
